+++ src/platform.ts
@@ -212,6 +212,21 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
       return;
     }
 
+    // Remove cached disallowed devices
+    if (
+      !this.isAllowedChannel(
+        accessory.context.deviceSerial,
+        accessory.context.channelId
+      )
+    ) {
+      this.removedAccessories.push(accessory);
+      this.log.warn(
+        "Removing accessory from cache (channel not on allow list):",
+        accessory.displayName
+      );
+      return;
+    }
+
     // Remove cached ignored devices
     if (
       this.isIgnoredChannel(
@@ -242,6 +257,7 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
       // Filter unsupported devices by serial range
       if (
         !serial.startsWith("ABB") && // free@home default
+        !serial.startsWith("0007") && // new free@home devices
         !serial.startsWith("E11") && // alarm services
         !serial.startsWith("7EB1") && // weather station
         !serial.startsWith("FFFF480") // Scenes
@@ -362,6 +378,14 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
       return false;
     }
 
+    // Filter allowed devices
+    if (!this.isAllowedChannel(serial, channelId)) {
+      this.log.debug(
+        `Ignored ${serial} (${channelId}): Channel is not listed on the allow list.`
+      );
+      return false;
+    }
+
     // Filter ignored devices
     if (this.isIgnoredChannel(serial, channelId)) {
       this.log.debug(
@@ -393,6 +417,7 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
     switch (functionID.toUpperCase() as FunctionID) {
       case FunctionID.FID_SWITCH_SENSOR:
       case FunctionID.FID_DIMMING_SENSOR:
+      case FunctionID.FID_DES_DOOR_RINGING_SENSOR:
         this.fahAccessories.set(
           `${serial}_${channelId}`,
           new SwitchSensorAccessory(this, accessory)
@@ -406,6 +431,7 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
         );
         return;
       case FunctionID.FID_ROOM_TEMPERATURE_CONTROLLER_MASTER_WITHOUT_FAN:
+      case FunctionID.FID_PANEL_ROOM_TEMPERATURE_CONTROLLER_MASTER_WITHOUT_FAN:
         this.fahAccessories.set(
           `${serial}_${channelId}`,
           new RoomTemperatureControllerAccessory(this, accessory)
@@ -465,6 +491,7 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
         return;
       case FunctionID.FID_WINDOW_DOOR_SENSOR:
       case FunctionID.FID_WINDOW_DOOR_POSITION_SENSOR:
+      case FunctionID.FID_WELCOME_IP_DOOR_OPEN_SENSOR:
         this.fahAccessories.set(
           `${serial}_${channelId}`,
           new ContactSensorAccessory(this, accessory)
@@ -505,6 +532,16 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
           `${serial} (${channelId}): Cannot configure accessory for FunctionID '${functionID}'!`
         );
     }
+  }
+
+  private isAllowedChannel(device: string, channel: string): boolean {
+    if (!this.config.allowedChannels) return true;
+
+    return (this.config.allowedChannels as Array<string>).some(
+      (e) =>
+        e.toUpperCase() === `${device.toUpperCase()}/*` ||
+        e.toUpperCase() === `${device.toUpperCase()}/${channel.toUpperCase()}`
+    );
   }
 
   private isIgnoredChannel(device: string, channel: string): boolean {