+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.17.0](https://github.com/pgerke/homebridge-freeathome-local-api/releases/tag/1.17.0) - 14.09.2025
+
+### Added
+
+- [#186](https://github.com/pgerke/homebridge-freeathome-local-api/issues/186): Support SmartTouch 10" Panel device
+- [#195](https://github.com/pgerke/homebridge-freeathome-local-api/issues/195): Support a channel / device allowlist
+
+### Changed
+
+- NO TICKET: Bump dependencies
+
 ## [1.16.0](https://github.com/pgerke/homebridge-freeathome-local-api/releases/tag/1.16.0) - 28.08.2025
 
 ### Fixed